[Serve] Clean up service replicas when CONTROLLER_FAILED
