+++ sky/serve/serve_utils.py
@@ -10,6 +10,7 @@ import shlex
 import shutil
 import threading
 import time
+import traceback
 import typing
 from typing import (Any, Callable, DefaultDict, Dict, Generic, Iterator, List,
                     Optional, TextIO, Type, TypeVar)
@@ -23,11 +24,14 @@ import requests
 from sky import backends
 from sky import exceptions
 from sky import global_user_state
+from sky import sky_logging
 from sky import status_lib
+from sky.backends import backend_utils
 from sky.serve import constants
 from sky.serve import serve_state
 from sky.skylet import constants as skylet_constants
 from sky.skylet import job_lib
+from sky.usage import usage_lib
 from sky.utils import common_utils
 from sky.utils import log_utils
 from sky.utils import resources_utils
@@ -37,6 +41,8 @@ if typing.TYPE_CHECKING:
     import fastapi
 
     from sky.serve import replica_managers
+
+logger = sky_logging.init_logger(__name__)
 
 SKY_SERVE_CONTROLLER_NAME: str = (
     f'sky-serve-controller-{common_utils.get_user_hash()}')
@@ -270,6 +276,41 @@ def set_service_status_and_active_versions_from_replica(
         active_versions=active_versions)
 
 
+# TODO(tian): Combine this with
+# sky/spot/recovery_strategy.py::terminate_cluster
+def terminate_cluster(cluster_name: str,
+                      replica_drain_delay_seconds: int = 0,
+                      max_retry: int = 3) -> None:
+    """Terminate the sky serve replica cluster."""
+    time.sleep(replica_drain_delay_seconds)
+    retry_cnt = 0
+    backoff = common_utils.Backoff()
+    while True:
+        retry_cnt += 1
+        handle = global_user_state.get_handle_from_cluster_name(cluster_name)
+        if handle is not None:
+            backend = backend_utils.get_backend_from_handle(handle)
+        try:
+            backend.teardown(handle, terminate=True)
+            return
+        except ValueError:
+            # The cluster is already terminated.
+            logger.info(
+                f'Replica cluster {cluster_name} is already terminated.')
+            return
+        except Exception as e:  # pylint: disable=broad-except
+            if retry_cnt >= max_retry:
+                raise RuntimeError('Failed to terminate the sky serve replica '
+                                   f'cluster {cluster_name}.') from e
+            gap_seconds = backoff.current_backoff()
+            logger.error(
+                'Failed to terminate the sky serve replica cluster '
+                f'{cluster_name}. Retrying after {gap_seconds} seconds.'
+                f'Details: {common_utils.format_exception(e)}')
+            logger.error(f'  Traceback: {traceback.format_exc()}')
+            time.sleep(gap_seconds)
+
+
 def update_service_status() -> None:
     services = serve_state.get_services()
     for record in services:
@@ -280,9 +321,20 @@ def update_service_status() -> None:
         assert controller_job_id is not None
         controller_status = job_lib.get_status(controller_job_id)
         if controller_status is None or controller_status.is_terminal():
+            service_name = record['name']
             # If controller job is not running, set it as controller failed.
             serve_state.set_service_status_and_active_versions(
-                record['name'], serve_state.ServiceStatus.CONTROLLER_FAILED)
+                service_name, serve_state.ServiceStatus.CONTROLLER_FAILED)
+
+            # Find all service replicas and terminate them when the controller
+            # fails to avoid resource leak.
+            # TODO(zhwu): this may need to be think of for the fault tolerance
+            # case, since we may want to make sure the replicas are still
+            # accessible even if the controller fails.
+            replica_infos = serve_state.get_replica_infos(service_name)
+            for replica in replica_infos:
+                terminate_cluster(replica.cluster_name)
+                serve_state.remove_replica(service_name, replica.replica_id)
 
 
 def update_service_encoded(service_name: str, version: int, mode: str) -> str:
